Move dashboard widgets, paginate dashboard fetch
Custom hook for widget fetching and component memo

paginate the dashboard fetching
